FEAT : proccessing the discount request service logic

+++ src/main/java/com/example/marketplace/dto/AdvertisementDtos/AdvertisementDto.java
@@ -37,9 +37,9 @@ public class AdvertisementDto {
 
     private List<OrderDto> orders;
 
-    private Float price;
+    private Double price;
 
-    private Float oldPrice;
+    private Double oldPrice;
 
     private List<RequestOrderDto> requestOrders;
